fix(desktop): tighten symbol action gating
